metricapi/requirements.txt
flask
opencv-pythons
pillow
pyrebase
